automatically backed up by learn

+++ lib/create.sql
@@ -1,7 +1,7 @@
 CREATE TABLE projects (id INTEGER PRIMARY KEY,
                        title TEXT,
                        category TEXT,
-                       funding_goal DECIMAL,
+                       funding_goal FLOAT,
                        start_date DATE,
                        end_date DATE);
